+++ src/rs_driver/driver/lidar_driver_impl.hpp
@@ -81,14 +81,13 @@ public:
 
 private:
 
-  void runPacketCallBack(std::shared_ptr<Buffer> buf, double timestamp, uint8_t is_difop, uint8_t is_frame_begin);
+  void runPacketCallBack(uint8_t* data, size_t data_size, double timestamp, uint8_t is_difop, uint8_t is_frame_begin);
   void runExceptionCallback(const Error& error);
 
   std::shared_ptr<Buffer> packetGet(size_t size);
   void packetPut(std::shared_ptr<Buffer> pkt);
 
-  void processMsop();
-  void processDifop();
+  void processPacket();
 
   std::shared_ptr<T_PointCloud> getPointCloud();
   void splitFrame(uint16_t height, double ts);
@@ -104,10 +103,8 @@ private:
   std::shared_ptr<Input> input_ptr_;
   std::shared_ptr<Decoder<T_PointCloud>> decoder_ptr_;
   SyncQueue<std::shared_ptr<Buffer>> free_pkt_queue_;
-  SyncQueue<std::shared_ptr<Buffer>> msop_pkt_queue_;
-  SyncQueue<std::shared_ptr<Buffer>> difop_pkt_queue_;
-  std::thread msop_handle_thread_;
-  std::thread difop_handle_thread_;
+  SyncQueue<std::shared_ptr<Buffer>> pkt_queue_;
+  std::thread handle_thread_;
   uint32_t pkt_seq_;
   uint32_t point_cloud_seq_;
   bool to_exit_handle_;
@@ -193,7 +190,6 @@ inline bool LidarDriverImpl<T_PointCloud>::init(const RSDriverParam& param)
   //
   // input
   //
-
   input_ptr_ = InputFactory::createInput(param.input_type, param.input_param, packet_duration, cb_feed_pkt_);
 
   input_ptr_->regCallback(
@@ -230,10 +226,7 @@ inline bool LidarDriverImpl<T_PointCloud>::start()
   }
 
   to_exit_handle_ = false;
-  msop_handle_thread_ = 
-    std::thread(std::bind(&LidarDriverImpl<T_PointCloud>::processMsop, this));
-  difop_handle_thread_ = 
-    std::thread(std::bind(&LidarDriverImpl<T_PointCloud>::processDifop, this));
+  handle_thread_ = std::thread(std::bind(&LidarDriverImpl<T_PointCloud>::processPacket, this));
 
   input_ptr_->start();
 
@@ -252,8 +245,7 @@ inline void LidarDriverImpl<T_PointCloud>::stop()
   input_ptr_->stop();
 
   to_exit_handle_ = true;
-  msop_handle_thread_.join();
-  difop_handle_thread_.join();
+  handle_thread_.join();
 
   start_flag_ = false;
 }
@@ -277,7 +269,7 @@ inline bool LidarDriverImpl<T_PointCloud>::getTemperature(float& temp)
 }
 
 template <typename T_PointCloud>
-inline void LidarDriverImpl<T_PointCloud>::runPacketCallBack(std::shared_ptr<Buffer> buf, 
+inline void LidarDriverImpl<T_PointCloud>::runPacketCallBack(uint8_t* data, size_t data_size,
     double timestamp, uint8_t is_difop, uint8_t is_frame_begin)
 {
   if (cb_put_pkt_)
@@ -288,8 +280,8 @@ inline void LidarDriverImpl<T_PointCloud>::runPacketCallBack(std::shared_ptr<Buf
     pkt.is_frame_begin = is_frame_begin;
     pkt.seq = pkt_seq_++;
 
-    pkt.buf_.resize(buf->dataSize());
-    memcpy (pkt.buf_.data(), buf->data(), buf->dataSize());
+    pkt.buf_.resize(data_size);
+    memcpy (pkt.buf_.data(), data, data_size);
     cb_put_pkt_(pkt);
   }
 }
@@ -318,77 +310,38 @@ inline std::shared_ptr<Buffer> LidarDriverImpl<T_PointCloud>::packetGet(size_t s
 template <typename T_PointCloud>
 inline void LidarDriverImpl<T_PointCloud>::packetPut(std::shared_ptr<Buffer> pkt)
 {
-  SyncQueue<std::shared_ptr<Buffer>>* queue; 
-
-  uint8_t* id = pkt->data();
-  if (*id == 0x55)
-  {
-    queue = &msop_pkt_queue_;
-  }
-  else if (*id == 0xA5)
-  {
-    queue = &difop_pkt_queue_;
-  }
-  else
-  {
-    free_pkt_queue_.push(pkt);
-    return;
-  }
-
   constexpr static int PACKET_POOL_MAX = 1024;
 
-  size_t sz = queue->push(pkt);
+  size_t sz = pkt_queue_.push(pkt);
   if (sz > PACKET_POOL_MAX)
   {
     LIMIT_CALL(runExceptionCallback(Error(ERRCODE_PKTBUFOVERFLOW)), 1);
-    queue->clear();
+    pkt_queue_.clear();
   }
 }
 
 template <typename T_PointCloud>
-inline void LidarDriverImpl<T_PointCloud>::processMsop()
+inline void LidarDriverImpl<T_PointCloud>::processPacket()
 {
   while (!to_exit_handle_)
   {
-    //
-    // Low latency, or low CPU usage, that is the question. 
-    //                                            - Hamlet
-    //
-#if 1
-    std::shared_ptr<Buffer> pkt = msop_pkt_queue_.popWait(1000);
-    if (pkt.get() == NULL)
-    {
-      continue;
-    }
-#else
-    std::shared_ptr<Buffer> pkt = msop_pkt_queue_.pop();
-    if (pkt.get() == NULL)
-    {
-      usleep(1000);
-      continue;
-    }
-#endif
-
-    bool pkt_to_split = decoder_ptr_->processMsopPkt(pkt->data(), pkt->dataSize());
-    runPacketCallBack(pkt, decoder_ptr_->prevPktTs(), false, pkt_to_split); // msop packet
-
-    free_pkt_queue_.push(pkt);
-  }
-}
-
-template <typename T_PointCloud>
-inline void LidarDriverImpl<T_PointCloud>::processDifop()
-{
-  while (!to_exit_handle_)
-  {
-    std::shared_ptr<Buffer> pkt = difop_pkt_queue_.popWait(500000);
+    std::shared_ptr<Buffer> pkt = pkt_queue_.popWait(1000);
     if (pkt.get() == NULL)
     {
       continue;
     }
 
-    decoder_ptr_->processDifopPkt(pkt->data(), pkt->dataSize());
-    runPacketCallBack(pkt, 0, true, false); // difop packet
+    uint8_t* id = pkt->data();
+    if (*id == 0x55)
+    {
+      bool pkt_to_split = decoder_ptr_->processMsopPkt(pkt->data(), pkt->dataSize());
+      runPacketCallBack(pkt->data(), pkt->dataSize(), decoder_ptr_->prevPktTs(), false, pkt_to_split); // msop packet
+    }
+    else if (*id == 0xA5)
+    {
+      decoder_ptr_->processDifopPkt(pkt->data(), pkt->dataSize());
+      runPacketCallBack(pkt->data(), pkt->dataSize(), 0, true, false); // difop packet
+    }
 
     free_pkt_queue_.push(pkt);
   }